@@ -1,7 +1,7 @@
 cmake_minimum_required (VERSION 3.15.0)
 
 project (KratosMultiphysics)
-set (CMAKE_CXX_STANDARD 17)
+set (CMAKE_CXX_STANDARD 20)
 set (CMAKE_CXX_STANDARD_REQUIRED ON)
 set (CMAKE_CXX_EXTENSIONS OFF)
 
@@ -169,6 +169,11 @@ ELSE(CMAKE_VERSION VERSION_LESS "3.16.0")
         ENDIF(${MSVC})
     ENDIF(CMAKE_UNITY_BUILD MATCHES ON)
 ENDIF(CMAKE_VERSION VERSION_LESS "3.16.0")
+
+# Add /utf-8 as a source file encoding flag for MSVC
+if(MSVC)
+    add_compile_options("/utf-8")
+endif()
 
 # Get subversion data. This is done automagically by the cmakes
 include (GenerateExportHeader)
