@@ -66,8 +66,11 @@ class GenericConstitutiveLawIntegratorDamage
     ///@name Type Definitions
     ///@{
 
+    /// The machine precision tolerance
+    static constexpr double tolerance = std::numeric_limits<double>::epsilon();
+
     /// The type of yield surface
-    typedef TYieldSurfaceType YieldSurfaceType;
+    using YieldSurfaceType = TYieldSurfaceType;
 
     /// The define the working dimension size, already defined in the yield surface
     static constexpr SizeType Dimension = YieldSurfaceType::Dimension;
@@ -321,6 +324,55 @@ class GenericConstitutiveLawIntegratorDamage
     static int Check(const Properties& rMaterialProperties)
     {
         KRATOS_ERROR_IF_NOT(rMaterialProperties.Has(SOFTENING_TYPE)) << "SOFTENING_TYPE is not a defined value" << std::endl;
+
+        const bool has_strain_vector = rMaterialProperties.Has(STRAIN_DAMAGE_CURVE);
+        const bool has_stress_vector = rMaterialProperties.Has(STRESS_DAMAGE_CURVE);
+
+        //Checks when using the CurveFittingDamage softening curve
+        KRATOS_ERROR_IF(has_strain_vector != has_stress_vector)
+            << "Bad definition of STRAIN_DAMAGE_CURVE / STRESS_DAMAGE_CURVE: both must be provided together. " << std::endl;
+
+        if (has_strain_vector && has_stress_vector) {
+
+            const Vector& r_strain_vector = rMaterialProperties[STRAIN_DAMAGE_CURVE];
+            const Vector& r_stress_vector = rMaterialProperties[STRESS_DAMAGE_CURVE];
+
+            KRATOS_ERROR_IF(r_strain_vector.size() != r_stress_vector.size())
+                << "Bad definition of STRAIN_DAMAGE_CURVE / STRESS_DAMAGE_CURVE: inconsistent dimensions. " << std::endl;
+
+            KRATOS_ERROR_IF(r_strain_vector.size() < 2)
+                << "Bad definition of STRAIN_DAMAGE_CURVE / STRESS_DAMAGE_CURVE: at least 2 points are required. Provided size: " << r_strain_vector.size() << std::endl;
+            KRATOS_ERROR_IF(r_strain_vector[0] <= tolerance)
+                << "Bad definition of STRAIN_DAMAGE_CURVE: first value cannot be 0. Provided STRAIN_DAMAGE_CURVE[0]=" << r_strain_vector[0] << std::endl;
+
+            for (SizeType i = 1; i < r_strain_vector.size(); ++i) {
+                KRATOS_ERROR_IF(r_strain_vector[i] <= r_strain_vector[i-1] + tolerance)
+                    << "Bad definition of STRAIN_DAMAGE_CURVE: it must be strictly increasing. " << std::endl;
+            }
+
+            double previous_stiffness = r_stress_vector[0] / r_strain_vector[0];
+            for (SizeType i = 1; i < r_strain_vector.size(); ++i) {
+                const double current_stiffness = r_stress_vector[i] / r_strain_vector[i];
+                KRATOS_ERROR_IF(current_stiffness > previous_stiffness + tolerance)
+                    << "Bad definition of STRAIN_DAMAGE_CURVE / STRESS_DAMAGE_CURVE: stiffness recovery detected (secant stiffness must be non-increasing). " << std::endl;
+                previous_stiffness = current_stiffness;
+            }
+
+            KRATOS_ERROR_IF_NOT(rMaterialProperties.Has(YIELD_STRESS))
+                << "Bad definition of material properties: YIELD_STRESS is required when using the by-points hardening/softening curve." << std::endl;
+
+            const double yield_stress = rMaterialProperties[YIELD_STRESS];
+            const double stress_tolerance = tolerance * yield_stress;
+
+            KRATOS_ERROR_IF(r_stress_vector[0] < yield_stress - stress_tolerance)
+                << "Bad definition of STRESS_DAMAGE_CURVE: first stress point is below YIELD_STRESS. Provided STRESS_DAMAGE_CURVE[0]=" << r_stress_vector[0]
+                << ", YIELD_STRESS=" << yield_stress << std::endl;
+
+            KRATOS_ERROR_IF(r_stress_vector[0] > yield_stress + stress_tolerance)
+                << "Bad definition of STRESS_DAMAGE_CURVE: first stress point is above YIELD_STRESS. Provided STRESS_DAMAGE_CURVE[0]=" << r_stress_vector[0]
+                << ", YIELD_STRESS=" << yield_stress << std::endl;
+        }
+
         return TYieldSurfaceType::Check(rMaterialProperties);
     }
 
