@@ -318,6 +318,21 @@ public:
         ) override;
 
     /**
+     * @brief This function provides the place to perform checks on the completeness of the input.
+     * @details It is designed to be called only once (or anyway, not often) typically at the beginning
+     * of the calculations, so to verify that nothing is missing from the input or that no common error is found.
+     * @param rMaterialProperties The properties of the material
+     * @param rElementGeometry The geometry of the element
+     * @param rCurrentProcessInfo The current process info instance
+     * @return 0 if OK, 1 otherwise
+     */
+    int Check(
+        const Properties& rMaterialProperties,
+        const GeometryType& rElementGeometry,
+        const ProcessInfo& rCurrentProcessInfo
+        ) const override;
+
+    /**
      * @brief If the CL requires to initialize the material response, called by the element in InitializeSolutionStep.
      */
     bool RequiresInitializeMaterialResponse() override
@@ -507,7 +522,7 @@ private:
         rSerializer.load("PreviousCycleTime", mPreviousCycleTime);
         rSerializer.load("Period", mPeriod);
         rSerializer.load("ReferenceDamage", mReferenceDamage);
-        rSerializer.save("PreviousCycleDamage", mPreviousCycleDamage);
+        rSerializer.load("PreviousCycleDamage", mPreviousCycleDamage);
         rSerializer.load("FirstCycleOfANewLoad", mFirstCycleOfANewLoad);
         rSerializer.load("CFactor", mCFactor);
     }
