@@ -17,8 +17,10 @@
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 
 #include <numbers>
+#include <string>
 
 using namespace std::numbers;
+using namespace std::string_literals;
 
 namespace Kratos::Testing
 {
@@ -155,7 +157,8 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesCorrectlyForConst
     })");
 
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     for (auto& node : model_part.Nodes()) {
         KRATOS_EXPECT_NEAR(node.FastGetSolutionStepValue(HYDRAULIC_HEAD), 1.0, 1e-6);
@@ -186,7 +189,8 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesCorrectlyForTwoCo
     })");
 
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     std::vector<double> expected_values = {1.0, 1.5, 1.5, 1.0, 2.0, 2.0};
     std::vector<double> actual_values;
@@ -223,7 +227,48 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesCorrectlyForLinea
          "list_of_variables"          : ["HYDRAULIC_HEAD"]
      })");
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
+
+    std::vector<double> expected_values = {-1, 0, 1, -1, -1, 1};
+    std::vector<double> actual_values;
+    actual_values.reserve(model_part.Nodes().size());
+    std::transform(model_part.Nodes().begin(), model_part.Nodes().end(), std::back_inserter(actual_values),
+                   [](const auto& node) { return node.FastGetSolutionStepValue(HYDRAULIC_HEAD); });
+
+    KRATOS_EXPECT_VECTOR_NEAR(actual_values, expected_values, 1e-6)
+}
+
+KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesCorrectlyForLinearFields_EvenIfUnrelatedEmptyModelPartsAreSupplied,
+                          KratosGeoMechanicsFastSuiteWithoutKernel)
+{
+    //   This test uses the following two-element system.
+    //   4------3------6
+    //   |  El1 |  El2 |
+    //   1------2------5
+    //
+    Model model;
+    auto& model_part = CreateModelPartWithTwoStubElements(model);
+    model.CreateModelPart("foo"s);
+    model.CreateModelPart("bar"s);
+
+    // Linear field in x between -1 and 1
+    dynamic_cast<StubElementForNodalExtrapolationTest&>(model_part.Elements()[1]).mIntegrationDoubleValues = {
+        -inv_sqrt3, inv_sqrt3, inv_sqrt3, -inv_sqrt3};
+
+    // Linear field in y between -1 and 1
+    dynamic_cast<StubElementForNodalExtrapolationTest&>(model_part.Elements()[2]).mIntegrationDoubleValues = {
+        -inv_sqrt3, -inv_sqrt3, inv_sqrt3, inv_sqrt3};
+
+    const auto                                         parameters = Parameters(R"(
+     {
+         "model_part_name_list" : ["MainModelPart", "foo", "bar"],
+         "echo_level"           : 0,
+         "list_of_variables"    : ["HYDRAULIC_HEAD"]
+     })");
+    GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     std::vector<double> expected_values = {-1, 0, 1, -1, -1, 1};
     std::vector<double> actual_values;
@@ -263,7 +308,8 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesMatrixCorrectlyFo
      })");
 
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     std::vector<Matrix> expected_values = {ScalarMatrix(3, 3, -1), ScalarMatrix(3, 3, 0),
                                            ScalarMatrix(3, 3, 1),  ScalarMatrix(3, 3, -1),
@@ -308,7 +354,8 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesVectorCorrectlyFo
      })");
 
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     std::vector<Vector> expected_values = {ScalarVector(6, -1), ScalarVector(6, 0),
                                            ScalarVector(6, 1),  ScalarVector(6, -1),
@@ -353,7 +400,8 @@ KRATOS_TEST_CASE_IN_SUITE(TestExtrapolationProcess_ExtrapolatesArrayCorrectlyFor
      })");
 
     GeoExtrapolateIntegrationPointValuesToNodesProcess process(model, parameters);
-    process.Execute();
+    process.ExecuteBeforeSolutionLoop();
+    process.ExecuteFinalizeSolutionStep();
 
     std::vector<Vector>              expected_values = {ScalarVector(3, -1), ScalarVector(3, 0),
                                                         ScalarVector(3, 1),  ScalarVector(3, -1),
