@@ -157,9 +157,10 @@ public:
     void Initialize(const ProcessInfo& rCurrentProcessInfo) override;
 
     /**
-     * Called at the beginning of each solution step
+     * Performs the particle to grid mapping
      */
-    void InitializeSolutionStep(const ProcessInfo& rCurrentProcessInfo) override;
+    void AddExplicitContribution(const ProcessInfo& rCurrentProcessInfo) override;
+    
 
 
     //************* COMPUTING  METHODS
