@@ -105,8 +105,12 @@ namespace Kratos
     KRATOS_CREATE_3D_VARIABLE_WITH_COMPONENTS( MPC_CONTACT_FORCE )
     KRATOS_CREATE_VARIABLE( int, MATERIAL_POINTS_PER_CONDITION )
     KRATOS_CREATE_VARIABLE( bool, IS_EQUAL_DISTRIBUTED )
+
+    // Hacks because of how we store internal variables. To be changed
     KRATOS_CREATE_VARIABLE( bool, MPC_CALCULATE_NODAL_REACTIONS )
     KRATOS_CREATE_VARIABLE( bool, MPC_CALCULATE_CONTACT_FORCE )
+    KRATOS_CREATE_VARIABLE( bool, CALCULATE_EXPLICIT_MP_STRESS)
+    KRATOS_CREATE_VARIABLE( bool, CALCULATE_MUSL_VELOCITY_FIELD)
 
     // Grid node variable
     KRATOS_CREATE_3D_VARIABLE_WITH_COMPONENTS( NODAL_MOMENTUM)
@@ -120,11 +124,9 @@ namespace Kratos
     KRATOS_CREATE_VARIABLE(double, PQMPM_SUBPOINT_MIN_VOLUME_FRACTION)
 
     // Explicit time integration variables
-    KRATOS_CREATE_VARIABLE(bool, CALCULATE_MUSL_VELOCITY_FIELD)
     KRATOS_CREATE_VARIABLE(bool, IS_EXPLICIT)
     KRATOS_CREATE_VARIABLE(bool, IS_EXPLICIT_CENTRAL_DIFFERENCE)
     KRATOS_CREATE_VARIABLE(int, EXPLICIT_STRESS_UPDATE_OPTION)
-    KRATOS_CREATE_VARIABLE(bool, CALCULATE_EXPLICIT_MP_STRESS)
     KRATOS_CREATE_VARIABLE(bool, EXPLICIT_MAP_GRID_TO_MP)
     KRATOS_CREATE_VARIABLE(bool, IS_FIX_EXPLICIT_MP_ON_GRID_EDGE)
 
