@@ -117,8 +117,14 @@ namespace Kratos
     KRATOS_DEFINE_3D_APPLICATION_VARIABLE_WITH_COMPONENTS(MPM_APPLICATION, MPC_CONTACT_FORCE )
     KRATOS_DEFINE_APPLICATION_VARIABLE( MPM_APPLICATION, int, MATERIAL_POINTS_PER_CONDITION )
     KRATOS_DEFINE_APPLICATION_VARIABLE( MPM_APPLICATION, bool, IS_EQUAL_DISTRIBUTED )
+
+    // Hacks because of how we store internal variables. To be changed
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, MPC_CALCULATE_NODAL_REACTIONS )
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, MPC_CALCULATE_CONTACT_FORCE )
+    KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, CALCULATE_EXPLICIT_MP_STRESS)
+    KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, CALCULATE_MUSL_VELOCITY_FIELD)
+
+
 
     // Essential Boundary Conditions
     KRATOS_DEFINE_APPLICATION_VARIABLE( MPM_APPLICATION, double, PENALTY_FACTOR )
@@ -147,11 +153,9 @@ namespace Kratos
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, double, PQMPM_SUBPOINT_MIN_VOLUME_FRACTION)
 
     // Explicit time integration variables
-    KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, CALCULATE_MUSL_VELOCITY_FIELD)
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, IS_EXPLICIT)
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, IS_EXPLICIT_CENTRAL_DIFFERENCE)
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, int, EXPLICIT_STRESS_UPDATE_OPTION)
-    KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, CALCULATE_EXPLICIT_MP_STRESS)
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, EXPLICIT_MAP_GRID_TO_MP)
     KRATOS_DEFINE_APPLICATION_VARIABLE(MPM_APPLICATION, bool, IS_FIX_EXPLICIT_MP_ON_GRID_EDGE)
 
