@@ -269,7 +269,7 @@ public:
      * @param ThisData Pointer to the data block associated with the node.
      * @param NewQueueSize The queue size for storing historical data (default = 1).
      */
-    Node(IndexType NewId, const double NewX, const double NewY, const double NewZ, 
+    Node(IndexType NewId, const double NewX, const double NewY, const double NewZ,
          VariablesList::Pointer pVariablesList, BlockType const * ThisData, SizeType NewQueueSize = 1)
          : BaseType(NewX, NewY, NewZ)
          , Flags()
@@ -380,17 +380,17 @@ public:
     {
         BaseType::operator=(rOther);
         Flags::operator =(rOther);
-    
+
         mNodalData = rOther.mNodalData;
-    
+
         // Deep copying the dofs
         for(typename DofsContainerType::const_iterator it_dof = rOther.mDofs.begin() ; it_dof != rOther.mDofs.end() ; it_dof++) {
             pAddDof(**it_dof);
         }
-    
+
         mData = rOther.mData;
         mInitialPosition = rOther.mInitialPosition;
-    
+
         return *this;
     }
 
@@ -403,7 +403,7 @@ public:
     bool operator==(const Node& rOther)
     {
         return Point::operator ==(rOther);
-    }    
+    }
 
     /**
      * @brief Accesses a solution step value for a given variable and step index.
@@ -462,7 +462,7 @@ public:
     double& operator[](IndexType ThisIndex)
     {
         return BaseType::operator[](ThisIndex);
-    }    
+    }
 
     /**
      * @brief Const accessor for node coordinates.
@@ -734,6 +734,48 @@ public:
         return mData.GetValue(rThisVariable);
     }
 
+
+    /**
+     * @brief Get the the data value if existing or create the value for a given @p rThisVariable.
+     * @details This method returns a reference to a value represented by @p rThisVariable from the
+     *          database. If the @p rThisVariable is not found, then a new value is created using
+     *          @p rThisVariable::Zero() method and then reference to new value is returned.
+     *
+     * @warning Multiple Emplace functions can be run concurrently OVER DIFFERENT DATABASES.
+     *          Concurrent Emplaces onto the same database ARE NOT THREADSAFE.
+     *
+     * @param rThisVariable     Variable representing the value.
+     * @return TDataType&       Reference to the value.
+     */
+    template<class TVariableType>
+    typename TVariableType::Type& Emplace(const TVariableType& rThisVariable)
+    {
+        return mData.Emplace(rThisVariable);
+    }
+
+    /**
+     * @brief Get the the data value if existing or create the value for a given @p rThisVariable.
+     * @details This method returns a reference to a value represented by @p rThisVariable from the
+     *          database. If the @p rThisVariable is not found, then a new value is created using
+     *          @p rInitValue and then reference to new value is returned.
+     *
+     *          In the case if @p rThisVariable is a component, then first a new value representing
+     *          the source variable is created with @p Zero() method, and then @p rInitValue
+     *          is used to initialize the component referred by @p rThisVariable.
+     *
+     * @warning Multiple Emplace functions can be run concurrently OVER DIFFERENT DATABASES.
+     *          Concurrent Emplaces onto the same database ARE NOT THREADSAFE.
+     *
+     * @param rThisVariable     Variable representing the value.
+     * @param rInitValue        Initialization value in case the @p rThisVariable is not found in the database.
+     * @return TDataType&       Reference to the value.
+     */
+    template<class TVariableType>
+    typename TVariableType::Type& Emplace(const TVariableType& rThisVariable, const typename TVariableType::Type& rInitValue)
+    {
+        return mData.Emplace(rThisVariable, rInitValue);
+    }
+
     /**
      * @brief Retrieves the value of a variable, checking both node data and solution step data.
      * @details If the variable is not present in the node's data, it fetches from the solution step data.
@@ -974,7 +1016,7 @@ public:
      * @param Y The Y coordinate of the new initial position.
      * @param Z The Z coordinate of the new initial position.
      */
-    void SetInitialPosition(double X,double Y, double Z)    
+    void SetInitialPosition(double X,double Y, double Z)
     {
         mInitialPosition.X() = X;
         mInitialPosition.Y() = Y;
@@ -989,7 +1031,7 @@ public:
     VariablesList::Pointer pGetVariablesList()
     {
         return SolutionStepData().pGetVariablesList();
-    }    
+    }
 
     /**
      * @brief Returns the pointer to the list of variables associated with the node (const version).
@@ -1149,7 +1191,7 @@ public:
 
     /**
      * @brief Adds a degree of freedom (DOF) to the node, returning the newly added DOF or the existing one if it already exists.
-     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists, 
+     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists,
      * it returns the existing DOF. Otherwise, it creates a new DOF for the given variable and adds it to the list.
      * The DOF list is then sorted to maintain order.
      * @tparam TVariableType The type of the variable associated with the degree of freedom.
@@ -1180,7 +1222,7 @@ public:
 
     /**
      * @brief Adds a degree of freedom (DOF) to the node, returning the newly added DOF or the existing one if it already exists.
-     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists, 
+     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists,
      * it updates the reaction and returns the existing DOF. If not, it creates a new DOF based on the provided `SourceDof`.
      * The DOF list is then sorted to maintain order.
      * @param SourceDof The DOF to be added or whose reaction should be updated.
@@ -1214,8 +1256,8 @@ public:
 
     /**
      * @brief Adds a degree of freedom (DOF) to the node, returning the newly added DOF or the existing one if it already exists, with a reaction.
-     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists, 
-     * it updates the reaction and returns the existing DOF. Otherwise, it creates a new DOF with both the variable 
+     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists,
+     * it updates the reaction and returns the existing DOF. Otherwise, it creates a new DOF with both the variable
      * and the reaction, and adds it to the list. The DOF list is then sorted to maintain order.
      * @tparam TVariableType The type of the variable associated with the degree of freedom.
      * @tparam TReactionType The type of the reaction associated with the degree of freedom.
@@ -1248,7 +1290,7 @@ public:
 
     /**
      * @brief Adds a degree of freedom (DOF) to the node, returning the newly added DOF or the existing one if it already exists.
-     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists, 
+     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists,
      * it returns the existing DOF. Otherwise, it creates a new DOF for the given variable and adds it to the list.
      * The DOF list is then sorted to maintain order.
      * @tparam TVariableType The type of the variable associated with the degree of freedom.
@@ -1279,8 +1321,8 @@ public:
 
     /**
      * @brief Adds a degree of freedom (DOF) to the node, returning the newly added DOF or the existing one if it already exists, with a reaction.
-     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists, 
-     * it updates the reaction and returns the existing DOF. Otherwise, it creates a new DOF with both the variable 
+     * @details This function checks if the DOF for the given variable already exists in the node's DOF list. If it exists,
+     * it updates the reaction and returns the existing DOF. Otherwise, it creates a new DOF with both the variable
      * and the reaction, and adds it to the list. The DOF list is then sorted to maintain order.
      * @tparam TVariableType The type of the variable associated with the degree of freedom.
      * @tparam TReactionType The type of the reaction associated with the degree of freedom.
@@ -1353,7 +1395,7 @@ public:
      * @brief Checks if the GeometricalObject is active
      * @return True by default, otherwise depending on the ACTIVE flag
      */
-    inline bool IsActive() const 
+    inline bool IsActive() const
     {
         return IsDefined(ACTIVE) ? Is(ACTIVE) : true;
     }
@@ -1434,7 +1476,7 @@ private:
     /// Storage for the dof of the node
     DofsContainerType  mDofs;
 
-    /// A container with data related to this node (non-historical variables) 
+    /// A container with data related to this node (non-historical variables)
     DataValueContainer mData;
 
     /// Initial Position of the node
@@ -1511,7 +1553,7 @@ private:
         rSerializer.save("Data", mData);
         rSerializer.save("Initial Position", mInitialPosition);
         rSerializer.save("Data", mDofs);
-    }    
+    }
 
     /**
      * The load operation which restores the database of the class
